Move `Error` and `RenderInfo` out of `html` module
